+++ bin/install_pepys.ps1
@@ -92,6 +92,8 @@ try {
     # Pepys Viewer shortcuts
     Make-Shortcut -ShortcutLocation ($startmenu_location + "Pepys\Pepys Viewer.lnk") -TargetPath ".\pepys_viewer.bat" -Icon $icon_string
     Make-Shortcut -ShortcutLocation ($startmenu_location + "Pepys\Pepys Viewer (training mode).lnk") -TargetPath ".\pepys_viewer_training.bat" -Icon $icon_string
+
+    Make-Shortcut -ShortcutLocation ($startmenu_location + "Pepys\Upgrade Pepys.lnk") -TargetPath ".\upgrade_pepys.bat" -Icon $icon_string
 }
 catch {
     Write-Output $_
@@ -109,7 +111,7 @@ try {
     if (!($env:Path -split ';' -contains $pepys_bin_path)) {
         [Environment]::SetEnvironmentVariable(
             "PATH",
-            [Environment]::GetEnvironmentVariable("PATH", [EnvironmentVariableTarget]::User) + ";" + $pepys_bin_path,
+            $pepys_bin_path + ";" + [Environment]::GetEnvironmentVariable("PATH", [EnvironmentVariableTarget]::User),
             [EnvironmentVariableTarget]::User)
     }
 }